+++ content/editor/input_repeater.go
@@ -2,49 +2,112 @@ package editor
 
 import (
 	"bytes"
-	"log"
-	"strings"
+	"fmt"
+	"reflect"
+	"strconv"
 )
 
-func InputRepeater(fieldName string, p interface{}, attrs map[string]string) []byte {
-	// find the field values in p to determine pre-filled inputs
-	fieldVals := ValueFromStructField(fieldName, p, nil).(string)
-	vals := strings.Split(fieldVals, "__ponzu")
+func InputRepeater(publicPath, fieldName string, p interface{}, attrs map[string]string, args *FieldArgs) []byte {
+	value := ValueByName(fieldName, p, args)
 
-	scope := TagNameFromStructField(fieldName, p, nil)
-	html := bytes.Buffer{}
-
-	_, err := html.WriteString(`<span class="__ponzu-repeat ` + scope + `">`)
-	if err != nil {
-		log.Println("Error writing HTML string to InputRepeater buffer")
-		return nil
+	if value.Kind() != reflect.Slice && value.Kind() != reflect.Array {
+		panic(fmt.Sprintf("Ponzu: Type '%s' for field '%s' not supported.", value.Type(), fieldName))
 	}
 
-	for i, val := range vals {
-		el := &Element{
-			TagName: "input",
-			Attrs:   attrs,
-			Name:    TagNameFromStructFieldMulti(fieldName, i, p),
-			Data:    val,
-			ViewBuf: &bytes.Buffer{},
-		}
+	scope := TagNameFromStructField(fieldName, p, args)
 
-		// only add the label to the first input in repeated list
-		if i == 0 {
-			el.Label = attrs["label"]
-		}
+	tmpl := `
+		<div class="control-block __ponzu-repeat ` + scope + `">
+			<label class="active">` + fieldName + `</label>
+	`
 
-		_, err := html.Write(DOMInputSelfClose(el))
-		if err != nil {
-			log.Println("Error writing DOMElementSelfClose to InputRepeater buffer")
-			return nil
-		}
-	}
-	_, err = html.WriteString(`</span>`)
-	if err != nil {
-		log.Println("Error writing HTML string to InputRepeater buffer")
-		return nil
+	positionalPlaceHolder := makePositionalPlaceholder(fieldName)
+	fieldArgs := &FieldArgs{
+		Parent:                 fmt.Sprintf("%s.%s", fieldName, positionalPlaceHolder),
+		PositionalPlaceHolders: []string{positionalPlaceHolder},
 	}
 
-	return append(html.Bytes(), RepeatController(fieldName, p, "input", ".input-field")...)
+	emptyType := makeEmptyType(p)
+	if value.IsZero() {
+		emptyType = p
+	}
+
+	if args != nil && args.Parent != "" {
+		fieldArgs.Parent = fmt.Sprintf("%s.%s", args.Parent, fieldArgs.Parent)
+		fieldArgs.PositionalPlaceHolders = append(
+			fieldArgs.PositionalPlaceHolders,
+			args.PositionalPlaceHolders...,
+		)
+
+		matches := ancestorIsFieldCollectionRegexp.FindStringSubmatch(args.Parent)
+		if len(matches) > 0 {
+			positionMatchIndex := ancestorIsFieldCollectionRegexp.SubexpIndex("Position")
+			if positionMatchIndex == -1 {
+				panic("Parent path is invalid")
+			}
+
+			fieldCollectionNameIndex := ancestorIsFieldCollectionRegexp.SubexpIndex("FieldCollectionName")
+			if fieldCollectionNameIndex == -1 {
+				panic("Parent path is invalid")
+			}
+
+			var err error
+			position, err := strconv.Atoi(matches[positionMatchIndex])
+			if err != nil {
+				panic(err)
+			}
+
+			emptyType, err = makeFieldCollectionAtPosition(p, matches[fieldCollectionNameIndex], args.TypeName, position)
+			if err != nil {
+				panic(err)
+			}
+		}
+	}
+
+	elementClass := fmt.Sprintf("%s-element", fieldName)
+	extendAttributeValue(attrs, "controlClass", elementClass)
+	entryTemplate := Input("", emptyType, attrs, fieldArgs)
+
+	script := &bytes.Buffer{}
+	scriptTmpl := makeScript("input_repeater")
+	if err := scriptTmpl.Execute(script, struct {
+		Template              string
+		NumItems              int
+		Scope                 string
+		CloneSelector         string
+		PositionalPlaceholder string
+		PublicPath            string
+		EntityName            string
+	}{
+		Template:              string(entryTemplate),
+		NumItems:              value.Len(),
+		Scope:                 scope,
+		CloneSelector:         fmt.Sprintf(".%s", elementClass),
+		PositionalPlaceholder: positionalPlaceHolder,
+		PublicPath:            publicPath,
+		EntityName:            fieldName,
+	}); err != nil {
+		panic(err)
+	}
+
+	for i := 0; i < value.Len(); i++ {
+		entryArgs := &FieldArgs{
+			Parent:                 fmt.Sprintf("%s.%d", fieldName, i),
+			PositionalPlaceHolders: []string{positionalPlaceHolder},
+		}
+
+		if args != nil && args.Parent != "" {
+			entryArgs.Parent = fmt.Sprintf("%s.%s", args.Parent, entryArgs.Parent)
+			entryArgs.PositionalPlaceHolders = append(
+				entryArgs.PositionalPlaceHolders,
+				args.PositionalPlaceHolders...,
+			)
+		}
+
+		fieldTemplate := Input("", p, attrs, entryArgs)
+		tmpl += string(fieldTemplate)
+	}
+
+	tmpl += `</div>`
+	return append([]byte(tmpl), script.Bytes()...)
 }