+++ CHANGELOG.md
@@ -1,5 +1,8 @@
+## v0.14.0
+- Reimplement `InputRepeater` using patterns borrowed from `NestedRepeater`
+
 ## v0.13.1
-- Fix Nested Repeater injecting child in wrong place with ':scope' selector without polyfill
+- Fix Nested Repeater injecting child in the wrong place with ':scope' selector without polyfill
 
 ## v0.13.0
 - Reference loader, load uncached objects